feat: Use approximate counts in data statistics [DHIS2-10865]
Introduces approximate counts of large tables in data statistics. Doing exact counts with select count(*) is no longer feasible on large databases for data values and events and makes the API request time out.

Adds counts for program instance and tracked entity instance entities.

Uses Long over Integer for object count numbers.

Adds counts for enrollment and TEI.

Changes the API controller default response format to JSON, as XML is broken for the DataSummary object due to numbers being used as map keys / XML elements.

(cherry picked from commit ee57f62237c3bf438445c2818d42053118568966)

+++ dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/statistics/StatisticsProvider.java
@@ -39,5 +39,5 @@ public interface StatisticsProvider
 {
     String ID = StatisticsProvider.class.getName();
 
-    Map<Objects, Integer> getObjectCounts();
+    Map<Objects, Long> getObjectCounts();
 }
